- applied amd64 patch
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@831 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/editor/ui_menus/editor_player_menu_allowed_buildings_menu.cc
@@ -132,7 +132,7 @@ void Editor_Player_Menu_Allowed_Buildings_Menu::clicked(int i) {
    
    
    // Remove from one list
-   int index=((int)source->get_selection());
+   long index=((long)source->get_selection());
    source->remove_entry(source->get_selection_index());
    source->sort();
